allow n to be really big
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@48240 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ raster/r.univar/globals.h
@@ -34,14 +34,14 @@ typedef struct
     unsigned int n_perc;
     double *perc;
     double sum_abs;
-    int n;
-    int size;
+    unsigned long n;
+    unsigned long size;
     DCELL *dcell_array;
     FCELL *fcell_array;
     CELL *cell_array;
     int map_type;
     void *nextp;
-    int n_alloc;
+    off_t n_alloc;
     int first;
 } univar_stat;
